Revert to old code due to use of bad sources

+++ fileio/mycat.c
@@ -2,13 +2,14 @@
 
 #define    BUFFSIZE    4096
 
-int main(void)
+int
+main(void)
 {
-    ssize_t n;
+    int n;
     char buf[BUFFSIZE];
 
     while ((n = read(STDIN_FILENO, buf, BUFFSIZE)) > 0) {
-        if (write(STDOUT_FILENO, buf, (size_t) n) != n) {
+        if (write(STDOUT_FILENO, buf, n) != n) {
             err_sys("write error");
         }
     }
